refactor(state): move RocksDB-specific shutdown to a new module
